Compute and plot integration assessment methods for Seurat object

Wrapper for application of single cell integration assessment metrics with default parameters. Takes in Seurat object, calculates integration assessment metric per cell for given grouping variables, and plots visualization of metrics. A single integration score per dataset can be reported as the median of the metric across all cells.

Usage

assess_integration(
  sobj,
  group.var,
  method = c("LISI", "MixingMetric"),
  reduction = c("umap", "pca", "tsne"),
  MM_flip = TRUE,
  title = NULL,
  savename = NULL,
  width = NULL,
  height = NULL
)

run_LISI(
  sobj,
  group.var,
  reduction = c("umap", "pca", "tsne"),
  title = NULL,
  savename = NULL,
  width = NULL,
  height = NULL
)

run_MixingMetric(
  sobj,
  group.var,
  MM_flip = TRUE,
  reduction = c("umap", "pca", "tsne"),
  title = NULL,
  savename = NULL,
  width = NULL,
  height = NULL
)

Arguments

sobj

Seurat object

group.var

character (vector); metadata colname to check level of integration

method

c("LISI", "MixingMetric"); integration assessment metric to apply

reduction

c("umap", "pca", "tsne"); dimensional reduction method

MM_flip

logical; use max.k - MixingMetric instead of MixingMetric only

title

string; plot title

savename

string; filepath to save results and figure under (no ext.)

width

numeric; plot width

height

numeric; plot height

Value

Dataframe of integration assessment values

Details

LISI: local inverse Simpson's index

Wrapper for immunogenomics/LISI from Harmony paper. Quantifies diversity of batches within local neighborhoods, defined by kNN graphs, using the inverse Simpson's index.

LISI = 1: neighborhood comprised of single group, not mixed

LISI = # of groups: neighborhood has equal number of cells from all datasets, perfect mixing

Seurat MixingMetric: local / group kNN ranking

Wrapper for Seurat::MixingMetric from Seurat v3 paper. Assesses local neighborhood composition for a cell and how well mixed that neighborhood is. If the neighborhood is well mixed, should contain at least some cells from each dataset. Can be more intuitive to have high MixingMetric mean good mixing, so range can be flipped with max.k - MixingMetric. Uses flipped scale (MM_flip = TRUE) by default.

MixingMetric = max.k, low max.k - MixingMetric: not mixed

Low MixingMetric, high max.k - MixingMetric: well mixed

Functions

run_LISI(): Compute and plot local inverse Simpson's index (LISI)

run_MixingMetric(): Compute and plot Seurat MixingMetric

See also