Plot GSEA pathway plots (waterfall + enrichment)

If subtitle is not specified, Kolmogorov-Smirnov enrichment p-value is calculated to check if the genes in the pathway separate from the genes not in the chosen pathway based on ranking by the desired rankcol.

Usage

plot_GSEA_pathway(
  sig,
  geneset,
  genecol = "gene",
  rankcol,
  rankcol_name = rankcol,
  hightolow = FALSE,
  label = length(geneset) < 20,
  legendpos = "none",
  lab_high = NULL,
  lab_low = NULL,
  hllab = "Highlight",
  title = NULL,
  subtitle = NULL,
  savename = NULL
)

Arguments

sig

dataframe; genecol + rankcol columns

geneset

vector; list of genes

genecol

string; colname of gene names in sig

rankcol

string; colname of values in sig

rankcol_name

string; descriptive name of values

hightolow

logical; T for high values on left, low on right

label

logical; T to label highlighted genes

legendpos

vector; value btwn 0-1 as legend coordinates (ggplot2 legend.position)

lab_high

string; description of high values

lab_low

string; description of low values

hllab

string; description of highlighted values

title

string; plot title

subtitle

string; plot subtitle; !! overwrites p-val_enrichment

savename

string; filepath to save png under

Value

grid of gene set enrichment waterfall above enrichment plot