Run RRHO analysis

Inference on the amount of agreement in two sorted lists using the Rank-Rank Hypergeometric Overlap test. Outputs RRHO results, hypermatrix heatmap, rank rank scatter, metric scatter.

Usage

run_RRHO(
  sig1,
  sig2,
  sig1_name,
  sig2_name,
  sig1_low = NA,
  sig1_high = NA,
  sig2_low = NA,
  sig2_high = NA,
  key = "gene",
  metric1 = "sign_log_p",
  metric2 = metric1,
  steps = NULL,
  savename = NULL,
  webtool = TRUE,
  BY = FALSE,
  hm_method = c("ggplot", "lattice"),
  palette = "Spectral",
  waterfall = TRUE,
  scatter = TRUE,
  plot_fmt = "png"
)

Arguments

sig1

string/dataframe; path or df for sig1, with cols key and metric1

sig2

string/dataframe; path or df for sig2, with cols key and metric2

sig1_name

string; description of sig1

sig2_name

string; description of sig2

sig1_low

string; description of low sig1 values

sig1_high

string; description of high sig1 values

sig2_low

string; description of low sig2 values

sig2_high

string; description of high sig2 values

key

string; colname of corresponding values btwn both sigs

metric1

string; colname of sig1 metric to rank by

metric2

string; colname of sig2 metric to rank by

steps

integer; RRHO step size

savename

string; filepath to save results under (no extension)

webtool

logical; T to output text formatted for Graeber RRHO webtool

BY

logical; T for Benjamini-Yekutieli FDR corrected pvalues

hm_method

string; make heatmap with ggplot geom_raster or lattice levelplot

palette

string; RColorBrewer continuous palette name

waterfall

logical; TRUE to include waterfall subplots (hm_method = "ggplot" only)

scatter

logical; TRUE to output metric & rank scatterplot

plot_fmt

string; file extension to save plot as

Value

RRHO results, as list of plots if scatter == TRUE or as ggplot RRHO heatmap object

Details

Arrows in unicode are \u2193 (down) and \u2191 (up) if you want to be cool in your low/high descriptions B)